refactor: extract results-kmb utils and stop chips hook (closes #50)
